✅ Reorganize test files structure and update test cases
